Replaced regular v8z with 3.28-zos branch.
Also hacked the configure file so that it would recognize if zos was
running on an s390 system. (ugly hack right now, plz fix). This was
done by hacking the function that gets predefined macros of compiler;
please modify this in the future so that it does the same thing that it
does with gcc, with xlc. (instead of just automatically picking s390 as
the cpu type if it detects zos as the os)